requirements.txt
apache_beam=2.4.0
Augmentor==0.2.2
backcall==0.1.0
bleach==2.1.3
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
future==0.16.0
graphviz==0.8.2
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipyleaflet==0.8.1
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mxnet-cu91==1.2.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numpy==1.14.2
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
pyproj==1.9.5.1
python-dateutil==2.7.2
pytz==2018.4
PyWavelets==0.5.2
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tifffile==0.14.0
tornado==5.0.2
tqdm==4.23.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.0
